+++ main.py
@@ -6,6 +6,7 @@ from binance.client import Client
 from signal_generators import SignalGenerator
 from base_strategy import MACDStateMachine
 from binance_account import BinanceAccount
+from models.run_setting import RunSetting
 
 ## API_KEY
 api_key = config.API_KEY
@@ -18,14 +19,28 @@ timestamp = startTime.timestamp()
 startTimeStr = startTime.strftime("%m/%d/%Y - %H:%M:%S")
 print(startTimeStr)
 
+# get run settings
+runSetting = RunSetting.query(limit=1, slug="default_bnbbusd")[0]
+
 ## run 
-cc  = 'bnbbusd' # must be lower case
-interval, lookback = kline_lookback_config.FIVE_MINUTE
+cc = runSetting.websocketSymbol
+
+interval, lookback = None, None
+if runSetting.tradeInterval == "ONE_MINUTE":
+    interval, lookback = kline_lookback_config.ONE_MINUTE
+elif runSetting.tradeInterval == "FIVE_MINUTE":
+    interval, lookback = kline_lookback_config.FIVE_MINUTE
+elif runSetting.tradeInterval == "FIFTEEN_MINUTE":
+    interval, lookback = kline_lookback_config.FIFTEEN_MINUTE
 
 socket = f"wss://stream.binance.com:9443/ws/{cc}@kline_{interval}"
+
+# TODO - gat from account
 client = Client(config.API_KEY, config.API_SECRET)
-symbol = "BNBBUSD"
-indicators = ['EMAS', 'MACD']
+
+symbol = runSetting.symbol
+
+indicators = runSetting.indicators
 
 # signal generator
 signal = SignalGenerator(client, symbol, lookback, interval)